requirements.txt
torch>=1.9.0
transformers>=4.20.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
# seqeval>=1.2.2
pytorch-crf>=0.7.2
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0
tabulate>=0.8.9
